requirements-dev.txt
pyspark[sql]==3.2.1
pandas==1.3.5
numpy==1.21.5
pytest==7.1.1
pytest-sugar==0.9.4
black==22.3.0
pylint==2.13.5
apache-airflow[celery]==2.2.5
isort==5.10.1